refactor(worker): extract the meeting-intelligence stage from pipeline.py into meeting_intelligence_stage.py
