+++ pyproject.toml
@@ -165,6 +165,12 @@ max-complexity = 10
 "backend/utils/transcript_utils.py" = ["C901", "PLR0912", "PLR0915"]
 "backend/worker/tasks/constants.py" = ["PLR0913"]
 "backend/worker/tasks/intelligence.py" = ["C901", "PLR0912", "PLR0913", "PLR0915"]
+"backend/worker/tasks/meeting_intelligence_stage.py" = [
+    "C901",
+    "PLR0912",
+    "PLR0913",
+    "PLR0915",
+]
 "backend/worker/tasks/pipeline.py" = ["C901", "PLR0912", "PLR0913", "PLR0915"]
 "backend/worker/tasks/system.py" = ["PLR0913"]
 # Tests are exempt from PLR0913/PLR0915 (fixtures and parametrised cases legitimately
